backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.3.4
ffmpeg-python==0.2.0
pydantic>=2.7.0
httpx==0.25.2
python-multipart==0.0.6
